Add new project files for Flask app
A new Flask project has been initialized with necessary configurations. This includes the creation of Python modules for the app and the database, as well as Flask and database configurations files. This commit also includes the models for the User and Product entities. Essential settings for the integrated development environment (IDE) are also part of this commit.